+++ src/program/ramsearch/MemScannerThread.cpp
@@ -31,6 +31,8 @@
 #include <thread>
 
 #define MEMORY_CHUNK_SIZE 1024*1024
+#define OUTPUT_CHUNK_SIZE 4096
+#define MAX_TYPE_SIZE 8
 
 MemScannerThread::MemScannerThread(MemScanner& ms, int br, int er, uintptr_t ba, uintptr_t ea, off_t mo, uint64_t mem) : memscanner(ms), beg_region(br), end_region(er), beg_address(ba), end_address(ea), memory_offset(mo), memory_size(mem), error(ENOERROR)
 {
@@ -125,8 +127,8 @@ void MemScannerThread::first_address_scan()
     processed_memory_size = 0;
 
     /* Save in files by batches */
-    uintptr_t batch_addresses[4096];
-    uint8_t batch_values[4096*8];
+    uintptr_t batch_addresses[OUTPUT_CHUNK_SIZE];
+    uint8_t batch_values[OUTPUT_CHUNK_SIZE*MAX_TYPE_SIZE];
     int batch_index = 0;
     
     /* Start searching from beg_address to end_address, which were split evenly
@@ -153,28 +155,32 @@ void MemScannerThread::first_address_scan()
         }
         
         /* Write data */
-        uint8_t chunk[4096];
+        uint8_t chunk[4096+MAX_TYPE_SIZE]; // extra size for unaligned search
         
         for (uintptr_t ca = cur_beg_addr; ca < cur_end_addr; ca += 4096) {
             processed_memory_size += 4096;
 
-            int readValues = MemAccess::read(chunk, reinterpret_cast<void*>(ca), 4096);
+            /* Compute how much extra data we need to read to account for unaligned
+             * search, which does not apply for the end of the region */
+            int extra_read = (4096+ca)<cur_end_addr ? memscanner.value_type_size-memscanner.alignment : 0;
+            
+            int readValues = MemAccess::read(chunk, reinterpret_cast<void*>(ca), 4096+extra_read);
             if (readValues < 0)
                 continue;
-            for (int v = 0; v < 4096; v += memscanner.value_type_size) {
+            for (int v = 0; v < readValues-(memscanner.value_type_size-memscanner.alignment); v += memscanner.alignment) {
                 if (CompareOperations::check_value(chunk+v)) {
                     batch_addresses[batch_index] = ca + v;
                     memcpy(batch_values+(batch_index*memscanner.value_type_size), chunk+v, memscanner.value_type_size);
                     batch_index++;
-                    if (batch_index == 4096) {
-                        afs.write((char*)batch_addresses, 4096*sizeof(uintptr_t));
-                        vfs.write((char*)batch_values, 4096*memscanner.value_type_size);
+                    if (batch_index == OUTPUT_CHUNK_SIZE) {
+                        afs.write((char*)batch_addresses, OUTPUT_CHUNK_SIZE*sizeof(uintptr_t));
+                        vfs.write((char*)batch_values, OUTPUT_CHUNK_SIZE*memscanner.value_type_size);
                         if (!afs || !vfs) {
                             error = EOUTPUT;
                             finished = true;
                             return;
                         }
-                        new_memory_size += 4096*memscanner.value_type_size;
+                        new_memory_size += OUTPUT_CHUNK_SIZE*memscanner.value_type_size;
                         batch_index = 0;
                     }
                 }
@@ -210,22 +216,22 @@ void MemScannerThread::next_scan_from_region()
     processed_memory_size = 0;
 
     std::vector<uint8_t> new_memory;
-    new_memory.resize(MEMORY_CHUNK_SIZE);
+    new_memory.resize(MEMORY_CHUNK_SIZE+memscanner.value_type_size-memscanner.alignment);
 
     /* If we compare from previous memory, read and process saved memory by
      * chunks and by region, because all threads access to the same file. */
     std::vector<char> old_memory;
     std::ifstream ivfs;
     if (memscanner.compare_type == CompareType::Previous) {
-        old_memory.resize(MEMORY_CHUNK_SIZE);
+        old_memory.resize(MEMORY_CHUNK_SIZE+memscanner.value_type_size-memscanner.alignment);
         
         ivfs.open(memscanner.values_path, std::ofstream::binary);
         ivfs.seekg(memory_offset);
     }
     
     /* Save in files by batches */
-    uintptr_t batch_addresses[4096];
-    uint8_t batch_values[4096*8];
+    uintptr_t batch_addresses[OUTPUT_CHUNK_SIZE];
+    uint8_t batch_values[OUTPUT_CHUNK_SIZE*MAX_TYPE_SIZE];
     int batch_index = 0;
 
     uintptr_t cur_beg_addr = beg_address;
@@ -251,25 +257,32 @@ void MemScannerThread::next_scan_from_region()
         
         /* Read chunks of memory */
         while (cur_beg_addr < cur_end_addr) {
-            int chunk_size = MEMORY_CHUNK_SIZE;
-            if ((cur_end_addr - cur_beg_addr) < chunk_size)
+            uint64_t chunk_size = MEMORY_CHUNK_SIZE;
+            uint64_t chunk_size_with_extra = chunk_size + memscanner.value_type_size-memscanner.alignment;
+            if ((cur_end_addr - cur_beg_addr) < chunk_size_with_extra) {
                 chunk_size = cur_end_addr - cur_beg_addr;
+                chunk_size_with_extra = chunk_size;
+            }
             
             processed_memory_size += chunk_size;
 
             if (memscanner.compare_type == CompareType::Previous) {
-                ivfs.read(old_memory.data(), chunk_size);
+                /* We need to seek each time, because we may need to read data
+                 * multiple times due to unaligned search */ 
+                ivfs.seekg(memory_offset);
+                ivfs.read(old_memory.data(), chunk_size_with_extra);
+                memory_offset += chunk_size;
             }
             
-            int readValues = MemAccess::read(new_memory.data(), reinterpret_cast<void*>(cur_beg_addr), chunk_size);
+            int readValues = MemAccess::read(new_memory.data(), reinterpret_cast<void*>(cur_beg_addr), chunk_size_with_extra);
             if (readValues < 0) {
                 std::cerr << "Cound not read game process at address " << cur_beg_addr << std::endl;
             }
-            if (readValues < chunk_size) {
+            if ((uint64_t)readValues < chunk_size_with_extra) {
                 std::cerr << "Did not read enough memory at address " << cur_beg_addr << std::endl;
             }
             
-            for (int v = 0; v < chunk_size; v += memscanner.value_type_size) {
+            for (unsigned int v = 0; v < chunk_size_with_extra-(memscanner.value_type_size-memscanner.alignment); v += memscanner.alignment) {
                 if (((memscanner.compare_type == CompareType::Previous) && 
                     CompareOperations::check_previous(&new_memory[v], &old_memory[v])) ||
                     ((memscanner.compare_type == CompareType::Value) && 
@@ -277,15 +290,15 @@ void MemScannerThread::next_scan_from_region()
                     batch_addresses[batch_index] = cur_beg_addr + v;
                     memcpy(batch_values+(batch_index*memscanner.value_type_size), &new_memory[v], memscanner.value_type_size);
                     batch_index++;
-                    if (batch_index == 4096) {
-                        afs.write((char*)batch_addresses, 4096*sizeof(uintptr_t));
-                        vfs.write((char*)batch_values, 4096*memscanner.value_type_size);
+                    if (batch_index == OUTPUT_CHUNK_SIZE) {
+                        afs.write((char*)batch_addresses, OUTPUT_CHUNK_SIZE*sizeof(uintptr_t));
+                        vfs.write((char*)batch_values, OUTPUT_CHUNK_SIZE*memscanner.value_type_size);
                         if (!afs || !vfs) {
                             error = EOUTPUT;
                             finished = true;
                             return;
                         }
-                        new_memory_size += 4096*memscanner.value_type_size;
+                        new_memory_size += OUTPUT_CHUNK_SIZE*memscanner.value_type_size;
                         batch_index = 0;
                     }
                 }
@@ -297,7 +310,7 @@ void MemScannerThread::next_scan_from_region()
                 }
             }
             
-            cur_beg_addr += chunk_size;            
+            cur_beg_addr += chunk_size;
         }
     }
     
@@ -325,11 +338,11 @@ void MemScannerThread::next_scan_from_address()
     int size_ratio = sizeof(uintptr_t)/memscanner.value_type_size;
 
     std::vector<uint8_t> new_memory;
-    new_memory.resize(4096);
+    new_memory.resize(4096+memscanner.value_type_size-memscanner.alignment);
 
     /* If we compare from previous memory, read and process saved memory by
      * chunks and by region, because all threads access to the same file. */
-    int max_chunk_size = MEMORY_CHUNK_SIZE;
+    uint64_t max_chunk_size = MEMORY_CHUNK_SIZE;
     if (memory_size < max_chunk_size)
         max_chunk_size = memory_size;
 
@@ -337,10 +350,9 @@ void MemScannerThread::next_scan_from_address()
 
     std::ifstream ivfs;
     if (memscanner.compare_type == CompareType::Previous) {
-        old_memory.resize(max_chunk_size);
+        old_memory.resize(max_chunk_size+memscanner.value_type_size-memscanner.alignment);
         
         ivfs.open(memscanner.values_path, std::ifstream::binary);
-        ivfs.seekg(memory_offset);
     }
 
     std::vector<uintptr_t> old_addresses;
@@ -349,24 +361,29 @@ void MemScannerThread::next_scan_from_address()
     std::ifstream iafs(memscanner.addresses_path, std::ifstream::binary);
     if (!iafs)
         std::cerr << "error: could not open file " << memscanner.addresses_path << std::endl;
-    iafs.seekg(memory_offset*size_ratio);
-    if (!iafs)
-        std::cerr << "error: could not seek to offset " << memory_offset*size_ratio << std::endl;
     
     /* Save in files by batches */
-    uintptr_t batch_addresses[4096];
-    uint8_t batch_values[4096*8];
+    uintptr_t batch_addresses[OUTPUT_CHUNK_SIZE];
+    uint8_t batch_values[OUTPUT_CHUNK_SIZE*MAX_TYPE_SIZE];
     int batch_index = 0;
 
     /* Read chunks of memory */
     uint64_t remaining_memory_size = memory_size;
     while (remaining_memory_size > 0) {
-        int chunk_size = MEMORY_CHUNK_SIZE;
-        if (remaining_memory_size < chunk_size)
+        uint64_t chunk_size = MEMORY_CHUNK_SIZE;
+        uint64_t chunk_size_with_extra = chunk_size + memscanner.value_type_size-memscanner.alignment;
+        if (remaining_memory_size < chunk_size_with_extra) {
             chunk_size = remaining_memory_size;
+            chunk_size_with_extra = chunk_size;
+        }
         
         if (memscanner.compare_type == CompareType::Previous) {
-            ivfs.read(old_memory.data(), chunk_size);
+            /* We need to seek each time, because we may need to read data
+             * multiple times due to unaligned search */ 
+            ivfs.seekg(memory_offset);
+            if (!ivfs)
+                std::cerr << "error: could not seek to offset " << memory_offset << std::endl;
+            ivfs.read(old_memory.data(), chunk_size_with_extra);
             if (!ivfs) {
                 std::cerr << "error: only " << ivfs.gcount() << " could be read from " << chunk_size << std::endl;
                 error = EINPUT;
@@ -374,13 +391,19 @@ void MemScannerThread::next_scan_from_address()
                 return;
             }
         }
-        iafs.read((char*)old_addresses.data(), chunk_size*size_ratio);
+        iafs.seekg(memory_offset*size_ratio);
+        if (!iafs)
+            std::cerr << "error: could not seek to offset " << memory_offset*size_ratio << std::endl;
+
+        iafs.read((char*)old_addresses.data(), chunk_size_with_extra*size_ratio);
         if (!iafs) {
             std::cerr << "error: only " << iafs.gcount() << " could be read from " << chunk_size << std::endl;
             error = EINPUT;
             finished = true;
             return;
         }
+        
+        memory_offset += chunk_size;
         
         int addr_beg_index = 0;
         int addr_end_index = chunk_size / memscanner.value_type_size;
@@ -432,15 +455,15 @@ void MemScannerThread::next_scan_from_address()
                     batch_addresses[batch_index] = addr;
                     memcpy(batch_values+(batch_index*memscanner.value_type_size), &new_memory[mem_index], memscanner.value_type_size);
                     batch_index++;
-                    if (batch_index == 4096) {
-                        afs.write((char*)batch_addresses, 4096*sizeof(uintptr_t));
-                        vfs.write((char*)batch_values, 4096*memscanner.value_type_size);
+                    if (batch_index == OUTPUT_CHUNK_SIZE) {
+                        afs.write((char*)batch_addresses, OUTPUT_CHUNK_SIZE*sizeof(uintptr_t));
+                        vfs.write((char*)batch_values, OUTPUT_CHUNK_SIZE*memscanner.value_type_size);
                         if (!afs || !vfs) {
                             error = EOUTPUT;
                             finished = true;
                             return;
                         }
-                        new_memory_size += 4096*memscanner.value_type_size;
+                        new_memory_size += OUTPUT_CHUNK_SIZE*memscanner.value_type_size;
                         batch_index = 0;
                     }
                 }
@@ -454,7 +477,6 @@ void MemScannerThread::next_scan_from_address()
             
             addr_beg_index = addr_cur_index;
         }
-        memory_offset += chunk_size;
         remaining_memory_size -= chunk_size;
     }